move BulkEditPinnedFilter-related models to models/filters.py
