+++ corehq/apps/data_cleaning/models/__init__.py
@@ -1,13 +1,11 @@
 import uuid
 
 from django.contrib.auth.models import User
-from django.contrib.postgres.fields import ArrayField
 from django.db import models, transaction
 
 from corehq.apps.data_cleaning.models.types import (
     BulkEditSessionType,
     DataType,
-    PinnedFilterType,
 )
 from dimagi.utils.chunked import chunked
 
@@ -442,62 +440,6 @@ class BulkEditSession(models.Model):
 
         self.result = result
         self.save(update_fields=['result'])
-
-
-class BulkEditPinnedFilterManager(models.Manager):
-    use_for_related_fields = True
-
-    def create_session_defaults(self, session):
-        default_types = {
-            BulkEditSessionType.CASE: PinnedFilterType.DEFAULT_FOR_CASE,
-        }.get(session.session_type)
-
-        if not default_types:
-            raise NotImplementedError(f"{session.session_type} default pinned filters not yet supported")
-
-        for index, filter_type in enumerate(default_types):
-            self.create(
-                session=session,
-                index=index,
-                filter_type=filter_type,
-            )
-
-    def apply_to_query(self, session, query):
-        for pinned_filter in session.pinned_filters.all():
-            query = pinned_filter.filter_query(query)
-        return query
-
-
-class BulkEditPinnedFilter(models.Model):
-    session = models.ForeignKey(BulkEditSession, related_name="pinned_filters", on_delete=models.CASCADE)
-    index = models.IntegerField(default=0)
-    filter_type = models.CharField(
-        max_length=11,
-        choices=PinnedFilterType.CHOICES,
-    )
-    value = ArrayField(
-        models.TextField(),
-        null=True,
-        blank=True,
-    )
-
-    objects = BulkEditPinnedFilterManager()
-
-    class Meta:
-        ordering = ["index"]
-
-    def get_report_filter_class(self):
-        from corehq.apps.data_cleaning.filters import (
-            CaseOwnersPinnedFilter,
-            CaseStatusPinnedFilter,
-        )
-        return {
-            PinnedFilterType.CASE_OWNERS: CaseOwnersPinnedFilter,
-            PinnedFilterType.CASE_STATUS: CaseStatusPinnedFilter,
-        }[self.filter_type]
-
-    def filter_query(self, query):
-        return self.get_report_filter_class().filter_query(query, self)
 
 
 class BulkEditColumnManager(models.Manager):
